feat: increase health stat after a walk

+++ wear/src/main/java/fr/croumy/bouge/presentation/models/companion/Stats.kt
@@ -3,5 +3,5 @@ package fr.croumy.bouge.presentation.models.companion
 data class Stats(
     val happiness: Int, // 0 TO 5
     val hungriness: Int, // 0 TO 5
-    val health: Int, // 0 TO 5
+    val health: Float, // 0 TO 5
 )
